bot/requirements.txt
aiogram==3.14.0
asyncpg==0.30.0
apscheduler==3.10.4
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
tzdata==2024.2
structlog==24.4.0
